refactor: Row, Column의 부모 Position 생성
